+++ cmdb/object_model_views.py
@@ -3,7 +3,7 @@ from django.shortcuts import render
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
 from cmdb.models import Vendor
-from cmdb.models import Model
+from cmdb.models import Model,HostPhysical
 import json
 import urllib
 import cmdb_log
@@ -103,6 +103,10 @@ def model_del(request):
     ids = data['id'].split(',')
     for del_id in ids:
         i = Model.objects.filter(id=del_id)
+        h = HostPhysical.objects.filter(model=del_id)
+        if len(h):
+            json_r = json.dumps({"result":"include hosts"})
+            return  HttpResponse(json_r)
         cmdb_log.log_deletion(request,i[0],i[0].Model_Name,data)
         i.delete()
     json_r = json.dumps({"result":"delete sucess"})
