+++ cmdb/object_os_views.py
@@ -3,6 +3,7 @@ from django.shortcuts import render
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
 from cmdb.models import OS
+from cmdb.models import HostPhysical
 import json
 import urllib
 import cmdb_log
@@ -82,6 +83,12 @@ def os_del(request):
     ids = data['id'].split(',')
     for del_id in ids:
         i = OS.objects.filter(id=del_id)
+        h = HostPhysical.objects.filter(os=del_id)
+        v = VirtualPhysical.objects.filter(os=del_id)
+        n = len(h)+len(v)
+        if n:
+            json_r = json.dumps({"result":"include hosts"})
+            return  HttpResponse(json_r)
         cmdb_log.log_deletion(request,i[0],i[0].OS_Name,data)
         i.delete()
     json_r = json.dumps({"result":"delete sucess"})
