+++ cmdb/object_rack_views.py
@@ -4,6 +4,7 @@ from django.http import HttpResponse
 from django.http import HttpResponseRedirect
 from cmdb.models import Rack 
 from cmdb.models import IDC
+from cmdb.models import HostPhysical
 import json
 import urllib
 import cmdb_log
@@ -105,6 +106,10 @@ def rack_del(request):
     ids = data['id'].split(',')
     for del_id in ids:
         i = Rack.objects.filter(id=del_id)
+        h = HostPhysical.objects.filter(rack=del_id)
+        if len(h):
+            json_r = json.dumps({"result":"include hosts"})
+            return  HttpResponse(json_r)
         cmdb_log.log_deletion(request,i[0],i[0].Rack_Name,data)
         i.delete()
     json_r = json.dumps({"result":"delete sucess"})
