+++ cmdb/object_raid_views.py
@@ -3,6 +3,7 @@ from django.shortcuts import render
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
 from cmdb.models import RAID 
+from cmdb.models import HostPhysical
 import json
 import urllib
 import cmdb_log
@@ -81,6 +82,10 @@ def raid_del(request):
     ids = data['id'].split(',')
     for del_id in ids:
         i = RAID.objects.filter(id=del_id)
+        h = HostPhysical.objects.filter(raid=del_id)
+        if len(h):
+            json_r = json.dumps({"result":"include hosts"})
+            return  HttpResponse(json_r)
         cmdb_log.log_deletion(request,i[0],i[0].RAID_Type,data)
         i.delete()
     json_r = json.dumps({"result":"delete sucess"})
