+++ cmdb/object_vendor_views.py
@@ -2,7 +2,7 @@
 from django.shortcuts import render
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
-from cmdb.models import Vendor 
+from cmdb.models import Vendor,HostPhysical 
 import json
 import urllib
 import cmdb_log
@@ -82,6 +82,10 @@ def vendor_del(request):
     ids = data['id'].split(',')
     for del_id in ids:
         i = Vendor.objects.filter(id=del_id)
+        h = HostPhysical.objects.filter(vendor=del_id)
+        if len(h):
+            json_r = json.dumps({"result":"include hosts"}) 
+            return  HttpResponse(json_r)
         cmdb_log.log_deletion(request,i[0],i[0].Vendor_Name,data)
         i.delete()
     json_r = json.dumps({"result":"delete sucess"})
