+++ cmdb/object_zone_views.py
@@ -4,6 +4,7 @@ from django.http import HttpResponse
 from django.http import HttpResponseRedirect
 from cmdb.models import Zone 
 from cmdb.models import IDC
+from cmdb.models import HostPhysical
 import json
 import urllib
 import cmdb_log
@@ -102,6 +103,12 @@ def zone_del(request):
     ids = data['id']
     for del_id in ids:
         i = Zone.objects.filter(id=del_id)
+        h = HostPhysical.objects.filter(zone=del_id)
+        v = VirtualPhysical.objects.filter(os=del_id)
+        n = len(h)+len(v)
+        if n:
+            json_r = json.dumps({"result":"include hosts"})
+            return  HttpResponse(json_r)
         cmdb_log.log_deletion(request,i[0],i[0].Zone_Name,data)
         i.delete()
     json_r = json.dumps({"result":"delete sucess"})
